add spring request validation, exceptions and exception handler; add checks that throw appropriate exceptions for all actions where are needed; change book entity to have access to related users; made other minor changes.

+++ library-api/src/main/java/com/my/library/service/impl/BookServiceImpl.java
@@ -1,10 +1,14 @@
 package com.my.library.service.impl;
 
 import com.my.library.dto.BookDto;
+import com.my.library.exception.BookAlreadyExistsException;
+import com.my.library.exception.BookNotFoundException;
+import com.my.library.exception.IssuedBookException;
 import com.my.library.mapper.BookMapper;
 import com.my.library.model.Book;
 import com.my.library.repository.BookRepository;
 import com.my.library.service.BookService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -13,17 +17,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@RequiredArgsConstructor
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
-
-    public BookServiceImpl(BookRepository bookRepository) {
-        this.bookRepository = bookRepository;
-    }
 
     @Transactional
     @Override
     public BookDto getById(long id) {
-        Book book = bookRepository.findById(id).orElse(null);
+        Book book = bookRepository.findById(id).orElseThrow(BookNotFoundException::new);
         return BookMapper.INSTANCE.mapBookDto(book);
     }
 
@@ -38,18 +39,23 @@ public class BookServiceImpl implements BookService {
     @Transactional
     @Override
     public BookDto create(BookDto bookDto) {
-        Book created = bookRepository.save(BookMapper.INSTANCE.mapBook(bookDto));
-        return BookMapper.INSTANCE.mapBookDto(created);
+        Book book = BookMapper.INSTANCE.mapBook(bookDto);
+        if (bookRepository.existsByAuthorAndName(book.getAuthor(), book.getName())) {
+            throw new BookAlreadyExistsException();
+        }
+        return BookMapper.INSTANCE.mapBookDto(bookRepository.save(book));
     }
 
     @Transactional
     @Override
-    public BookDto update(BookDto book) {
-        Book persisted = bookRepository.findById(book.getId()).orElse(null);
-        if (persisted == null) {
-            throw new RuntimeException("Book was not found!");
+    public BookDto update(BookDto bookDto) {
+        Book updating = BookMapper.INSTANCE.mapBook(bookDto);
+        Book persisted = bookRepository.findById(updating.getId()).orElseThrow(BookNotFoundException::new);
+        String updatingAuthor = updating.getAuthor() == null ? persisted.getAuthor() : updating.getAuthor();
+        String updatingName = updating.getName() == null ? persisted.getName() : updating.getName();
+        if (bookRepository.existsByAuthorAndNameAndIdIsNot(updatingAuthor, updatingName, updating.getId())) {
+            throw new BookAlreadyExistsException();
         }
-        Book updating = BookMapper.INSTANCE.mapBook(book);
         BookMapper.INSTANCE.mapPresentFields(persisted, updating);
         return BookMapper.INSTANCE.mapBookDto(bookRepository.save(persisted));
     }
@@ -57,6 +63,10 @@ public class BookServiceImpl implements BookService {
     @Transactional
     @Override
     public void deleteById(long id) {
+        Book book = bookRepository.findById(id).orElseThrow(BookNotFoundException::new);
+        if (!book.getUsers().isEmpty()) {
+            throw new IssuedBookException();
+        }
         bookRepository.deleteById(id);
     }
 }
